requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pymongo==4.6.2
pydantic==2.6.3
pydantic-settings==2.2.1
openai
python-dotenv==1.0.1
mongomock==4.1.2
streamlit
